+++ autotransform/change/__init__.py
@@ -1,0 +1,12 @@
+# AutoTransform
+# Large scale, component based code modification library
+#
+# Licensed under the MIT License <http://opensource.org/licenses/MIT>
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2022-present Nathan Rockenbach <http://github.com/nathro>
+
+# @black_format
+
+"""A change represents a submission from a run of AutoTransform on a particular Batch. They
+are used for managing submissions to code review/source control systems. A pull request is
+an example of a potential change."""