+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Version 0.2.0 represents a production ready beta of AutoTransform. It includes a
  - Runner class added to provide an API for invoking runs
  - GithubRunner class added to use Github workflows for remote runs
 
+ - Change class added to provide an API for handling submissions to source control/code review
+
 #### Scripts
 
  - Run script has been migrated to the main script